#5 - more type documentation
