+++ src/tree/types.js
@@ -1,0 +1,17 @@
+import { AbstractNode } from "./models.js";
+
+/**
+ * A function to sort DataCards
+ * @callback NodeComparator
+ * @param {AbstractNode} a
+ * @param {AbstractNode} b
+ * @returns {1 | 0 | -1}
+ */
+
+/**
+ * @callback StringCleaner
+ * @param {string} input
+ * @returns {string}
+ */
+
+export const types = {}
